requirements.txt
streamlit==1.29.0
anthropic==0.25.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
spacy==3.7.2
nltk==3.8.1
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
```
- Paste and Save

### File 9: `.env`
- New File in main folder
- Name: `.env`
- Type this:
```
ANTHROPIC_API_KEY=your_key_here